fix keystore LOG [-Wformat=]  warning
keystore/keymaster_enforcement.cpp:467:15: warning: format ‘%u’ expects argument of type ‘unsigned int’, but argument 4 has type ‘android::hardware::keymaster::V4_0::HardwareAuthenticatorType’ [-Wformat=]
keystore/keystore_aidl_hidl_marshalling_utils.cpp:128:15: warning: format ‘%d’ expects argument of type ‘int’, but argument 4 has type ‘android::hardware::keymaster::V4_0::Tag’ [-Wformat=]
keystore/confirmation_manager.cpp:92:15: warning: format ‘%d’ expects argument of type ‘int’, but argument 4 has type ‘keystore::ConfirmationResponseCode’ {aka ‘android::hardware::confirmationui::V1_0::ResponseCode’} [-Wformat=]
keystore/KeyStore.cpp:187:23: warning: format ‘%d’ expects argument of type ‘int’, but argument 4 has type ‘ResponseCode’ [-Wformat=]
warning: format '%d' expects argument of type 'int', but argument 4 has type 'int32_t' {aka 'long int'} [-Wformat=]

+++ keystore/keystore_aidl_hidl_marshalling_utils.cpp
@@ -92,7 +92,7 @@ NullOr<KeyParameter> readKeyParameterFromParcel(const android::Parcel& in) {
         result.blob = readKeymasterBlob(in);  // byte array
         break;
     default:
-        ALOGE("Unsupported KeyParameter tag %d", tag);
+        ALOGE("Unsupported KeyParameter tag %" PRId32, static_cast<int32_t>(tag));
         return {};
     }
     return result;
@@ -125,7 +125,7 @@ android::status_t writeKeyParameterToParcel(const KeyParameter& param, android::
         rc = writeKeymasterBlob(param.blob, out);
         break;
     default:
-        ALOGE("Failed to write KeyParameter: Unsupported tag %d", param.tag);
+        ALOGE("Failed to write KeyParameter: Unsupported tag %" PRId32, static_cast<uint32_t>(param.tag));
         rc = android::BAD_VALUE;
         break;
     }
